pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "financial_monitoring_engine"
version = "0.1.0"
description = "Risk engine and financial monitoring"
authors = [{name = "Fernando", email = "pickfer3@gmail.com"}]
dependencies = [
    "pandas",
    "yfinance",
    "numpy",
    "PyYAML" 
]

[tool.setuptools.packages.find]
where = ["src"]